pyproject.toml
[project]
name = "magic-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.3"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
